requirements.txt
Django==1.9.7
django-compressor==1.5
django-debug-toolbar==1.4
django-extensions==1.6.7
ipython==4.2.0
psycopg2==2.6.1
six==1.10.0
sqlparse==0.1.19
Pillow==3.2.0
easy-thumbnails==2.3
django-singletons==0.1.6
djangorestframework==3.3.3
django-cors-headers==1.1.0
